refactor(app): batch3 — extract quick-switch + navigation hooks
